+++ Caddisfly/src/main/java/com/ternup/caddisfly/app/GlobalsBase.java
@@ -78,7 +78,7 @@ public class GlobalsBase {
     public static final int SAMPLING_COUNT_DEFAULT = 5;
 
     // width and height of cropped image
-    public static final int SAMPLE_CROP_LENGTH_DEFAULT = 150;
+    public static final int SAMPLE_CROP_LENGTH_DEFAULT = 100;
 
     public static final String RESULT_VALUE_KEY = "resultValue";