+++ Caddisfly/src/main/java/com/ternup/caddisfly/util/DataHelper.java
@@ -16,10 +16,15 @@
 
 package com.ternup.caddisfly.util;
 
+import android.content.Context;
+import android.graphics.Color;
+import android.os.Bundle;
+
 import com.ternup.caddisfly.R;
 import com.ternup.caddisfly.app.Globals;
 
-import android.content.Context;
+import java.math.BigDecimal;
+import java.math.RoundingMode;
 
 public class DataHelper {
 
@@ -28,7 +33,7 @@ public class DataHelper {
             case Globals.FLUORIDE_INDEX:
                 return context.getString(R.string.fluoride);
             case Globals.FLUORIDE_2_INDEX:
-                return context.getString(R.string.fluoride2);
+                return context.getString(R.string.fluoride);
             case Globals.PH_INDEX:
                 return context.getString(R.string.pH);
             case Globals.BACTERIA_INDEX:
@@ -77,4 +82,111 @@ public class DataHelper {
                 return context.getString(R.string.error);
         }
     }
+
+    public static void saveResultToPreferences(Context context, int testType, long id) {
+        int samplingCount = PreferencesUtils
+                .getInt(context, R.string.samplingCountKey, Globals.SAMPLING_COUNT_DEFAULT);
+        for (int i = 0; i < samplingCount; i++) {
+            String key = String.format(context.getString(R.string.samplingIndexKey), i);
+            double tempResult = PreferencesUtils.getDouble(context, key);
+            PreferencesUtils.setDouble(context,
+                    String.format(context.getString(R.string.resultValueKey), testType, id, i),
+                    tempResult);
+
+            key = String.format(context.getString(R.string.samplingColorIndexKey), i);
+            int tempColor = PreferencesUtils.getInt(context, key, -1);
+            PreferencesUtils.setInt(context,
+                    String.format(context.getString(R.string.resultColorKey), testType, id, i),
+                    tempColor);
+
+            key = String.format(context.getString(R.string.samplingQualityIndexKey), i);
+            int tempQuality = PreferencesUtils.getInt(context, key, -1);
+            PreferencesUtils.setInt(context,
+                    String.format(context.getString(R.string.resultQualityKey), testType, id, i),
+                    tempQuality);
+        }
+
+    }
+
+    public static void saveResult(Context context, int testType, int id, int position, double resultValue, int resultColor, int quality) {
+
+        PreferencesUtils.setDouble(context,
+                String.format(context.getString(R.string.resultValueKey), testType, id, position),
+                resultValue);
+
+        PreferencesUtils.setInt(context,
+                String.format(context.getString(R.string.resultColorKey), testType, id, position),
+                resultColor);
+
+        PreferencesUtils.setInt(context,
+                String.format(context.getString(R.string.resultQualityKey), testType, id, position),
+                quality);
+    }
+
+    public static double getAverageResult(Context context, Bundle bundle) {
+
+        double result = 0;
+
+        int samplingCount = PreferencesUtils
+                .getInt(context, R.string.samplingCountKey, Globals.SAMPLING_COUNT_DEFAULT);
+        int counter = 0;
+        double commonResult = 0;
+        double[] results = new double[samplingCount];
+        int[] colors = new int[samplingCount];
+        for (int i = 0; i < samplingCount; i++) {
+            String key = String.format(context.getString(R.string.samplingIndexKey), i);
+            results[i] = PreferencesUtils.getDouble(context, key);
+            key = String.format(context.getString(R.string.samplingColorIndexKey), i);
+            colors[i] = PreferencesUtils.getInt(context, key, -1);
+            commonResult = ColorUtils.mostFrequent(results);
+        }
+
+        int red = 0;
+        int green = 0;
+        int blue = 0;
+        for (int i = 0; i < results.length; i++) {
+            if (results[i] >= 0 && colors[i] != -1) {
+                if (Math.abs(results[i] - commonResult) < 0.5) {
+                    counter++;
+                    result += results[i];
+                    red += Color.red(colors[i]);
+                    green += Color.green(colors[i]);
+                    blue += Color.blue(colors[i]);
+                }
+            }
+        }
+
+        if (counter > 0) {
+            result = round(result / counter, 2);
+            bundle.putDouble(Globals.RESULT_VALUE_KEY, result);
+            bundle.putInt(Globals.RESULT_COLOR_KEY,
+                    Color.rgb(red / counter, green / counter, blue / counter));
+        } else {
+            result = -1;
+        }
+
+        return result;
+    }
+
+    //Ref: http://stackoverflow.com/questions/2808535/round-a-double-to-2-decimal-places
+    public static double round(double value, int places) {
+        if (places < 0) throw new IllegalArgumentException();
+
+        BigDecimal bd = new BigDecimal(value);
+        bd = bd.setScale(places, RoundingMode.HALF_UP);
+        return bd.doubleValue();
+    }
+
+    public static void saveTempResult(Context context, double resultValue, int resultColor, int quality) {
+
+        int samplingCount = PreferencesUtils.getInt(context, R.string.currentSamplingCountKey, 0);
+        String key = String.format(context.getString(R.string.samplingIndexKey), samplingCount);
+        PreferencesUtils.setDouble(context, key, resultValue);
+        key = String.format(context.getString(R.string.samplingColorIndexKey), samplingCount);
+        PreferencesUtils.setInt(context, key, resultColor);
+        key = String.format(context.getString(R.string.samplingQualityIndexKey), samplingCount);
+        PreferencesUtils.setInt(context, key, quality);
+    }
+
+
 }